frontend quantity int -> double + db fix

+++ javafx/src/main/java/com/example/javafx/Material.java
@@ -3,10 +3,10 @@ package com.example.javafx;
 public class Material {
     private Integer id;
     private String name;
-    private Integer quantity;
+    private Double quantity;
     private Double price;
 
-    public Material(String name, Integer quantity, Double price) {
+    public Material(String name, Double quantity, Double price) {
         this.name = name;
         this.quantity = quantity;
         this.price = price;
@@ -23,7 +23,7 @@ public class Material {
         this.id = id;
     }
 
-    public Material(Integer id, String name, Integer quantity, Double price) {
+    public Material(Integer id, String name, Double quantity, Double price) {
         this.id = id;
         this.name = name;
         this.quantity = quantity;
@@ -34,7 +34,7 @@ public class Material {
         return name;
     }
 
-    public Integer getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
 
@@ -46,7 +46,7 @@ public class Material {
         this.name = name;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Double quantity) {
         this.quantity = quantity;
     }
